--- frontend/mockdata/mockConsultants.ts
@@ -1,12 +1,0 @@
-import { Consultant } from "@/types";
-
-export const MockConsultants: Consultant[] = [
-  {
-    id: "id",
-    name: "Test Consultant",
-    email: "test@company.io",
-    competences: ["Frontend"],
-    department: "My Department",
-    bookings: [{ year: 2023, weekNumber: 10, bookedHours: 10 }],
-  },
-];
