--- frontend/mockdata/mockDepartments.ts
@@ -1,8 +1,0 @@
-import { Department, Consultant } from "@/types";
-
-export const MockDepartments: Department[] = [
-  {
-    id: "myDepartment",
-    name: "My Department",
-  },
-];
